Add a checkbox to allow analysis while printing.
This fixes https://github.com/eyal0/OctoPrint-PrintTimeGenius/issues/53

+++ octoprint_PrintTimeGenius/__init__.py
@@ -134,6 +134,11 @@ class GeniusAnalysisQueue(GcodeAnalysisQueue):
     super(GeniusAnalysisQueue, self).__init__(finished_callback)
     self._plugin = plugin
 
+  def _do_abort(self, reenqueue=True):
+    super(GeniusAnalysisQueue, self)._do_abort(reenqueue)
+    if self._plugin._settings.get(["allowAnalysisWhilePrinting"]):
+      self._plugin._logger.info("Abort requested but will be ignored due to settings.")
+
   def compensate_analysis(self, analysis):
     logger = self._plugin._logger
     """Compensate for the analyzed print time by looking at previous statistics of
@@ -191,6 +196,7 @@ class GeniusAnalysisQueue(GcodeAnalysisQueue):
       logger.warning("Failed to compensate", exc_info=e)
 
   def _do_analysis(self, high_priority=False):
+    self._aborted = False
     logger = self._plugin._logger
     results = {}
     if self._plugin._settings.get(["enableOctoPrintAnalyzer"]):
@@ -219,7 +225,7 @@ class GeniusAnalysisQueue(GcodeAnalysisQueue):
         else:
           psutil.Process(popen.pid).nice(19)
         while popen.poll() is None:
-          if self._aborted:
+          if self._aborted and not self._plugin._settings.get(["allowAnalysisWhilePrinting"]):
             popen.terminate()
             raise AnalysisAborted(reenqueue=self._reenqueue)
           time.sleep(0.5)
@@ -290,6 +296,7 @@ class PrintTimeGeniusPlugin(octoprint.plugin.SettingsPlugin,
             for x in built_in_analyzers],
         "exactDurations": True,
         "enableOctoPrintAnalyzer": False,
+        "allowAnalysisWhilePrinting": False,
         "print_history": [],
         "printer_config": []
     }
@@ -331,6 +338,11 @@ class PrintTimeGeniusPlugin(octoprint.plugin.SettingsPlugin,
   @octoprint.plugin.BlueprintPlugin.route("/analyse/<origin>/<path:path>", methods=["GET"])
   def analyze_file(self, origin, path):
     """Add a file to the analysis queue."""
+    if not self._settings.get(["allowAnalysisWhilePrinting"]) and self._printer.is_printing():
+      self._file_manager._analysis_queue.pause()
+    else:
+      self._file_manager._analysis_queue.resume()
+
     queue_entry = self._file_manager._analysis_queue_entry(origin, path)
     if queue_entry is None:
       return ""
